Fix weird oauth format + time blocks.
